requirements.txt
opencv-python>=4.8.0
pyzbar>=0.1.9
Pillow>=10.0.0
pyautogui>=0.9.50
numpy>=1.24.0
pyinstaller>=6.0.0
qrcode[pil]>=7.0.0